Implement UDMF clipmidtex

+++ src/r_defs.h
@@ -579,8 +579,7 @@ typedef struct line_s
 	fixed_t dx, dy; // Precalculated v2 - v1 for side checking.
 	angle_t angle; // Precalculated angle between dx and dy
 
-	// Animation related.
-	INT16 flags;
+	UINT32 flags;
 	INT16 special;
 	taglist_t tags;
 	INT32 args[NUMLINEARGS];
